+++ tripal_chado/src/Task/ChadoPreparer.php
@@ -204,204 +204,41 @@ class ChadoPreparer extends ChadoTaskBase {
    * Loads ontologies necessary for creation of default Tripal content types.
    */
   protected function loadOntologies() {
-    \Drupal::service('tripal.tripalVocab.manager')->addVocabulary([
-      'idspace' => 'NCIT',
-      'namespace' => 'ncit',
-      'name' => 'NCI Thesaurus OBO Edition',
-      'description' => 'The NCIt OBO Edition project aims to increase integration of the NCIt with OBO Library ontologies.',
-      'url' => 'http://purl.obolibrary.org/obo/ncit.owl',
-      'urlprefix' => ' http://purl.obolibrary.org/obo/{db}_{accession}',
-    ]);
 
-    \Drupal::service('tripal.tripalTerm.manager')->addTerm([
-      'accession' => 'C25693',
-      'name' => 'Subgroup',
-      'vocabulary' => [
-        'namespace' => 'ncit',
-        'idspace' => 'NCIT',
-      ],
-      'definition' => 'A subdivision of a larger group with members often exhibiting similar characteristics. [ NCI ]',
-    ]);
-
-    \Drupal::service('tripal.tripalVocab.manager')->addVocabulary([
-      'idspace' => 'rdfs',
-      'namespace' => 'rdfs',
-      'name' => 'Resource Description Framework Schema',
-      'description' => 'RDF Schema provides a data-modelling vocabulary for RDF data.',
-      'url' => 'https://www.w3.org/TR/rdf-schema/',
-      'urlprefix' => 'http://www.w3.org/2000/01/rdf-schema#{accession}',
-    ]);
-
-    \Drupal::service('tripal.tripalTerm.manager')->addTerm([
-      'accession' => 'comment',
-      'name' => 'comment',
-      'vocabulary' => [
-        'namespace' => 'rdfs',
-        'idspace' => 'rdfs',
-      ],
-      'definition' => 'A human-readable description of a resource\'s name.',
-    ]);
-
-    // TODO:
-    // T3 loads many terms we need for default content types through the OBO
-    // Importer. As of April 19, 2021 the OBO importer has not yet been migrated
-    // to T4 and therefore cannot be used. As a result, I will be adding a few
-    // terms here individually using the vocab and term managers. In the
-    // future, this should be replaced with calls to the OBO importer so that
-    // the terms can be imported automatically.
-    $terms = [];
-
-    // Organism.
-    $terms['SO:0100026'] = [
-      'accession' => '0100026',
-      'name' => 'organism',
-      'definition' => 'A material entity that is an individual living system, such as animal, plant, bacteria or virus, that is capable of replicating or reproducing, growth and maintenance in the right environment. An organism may be unicellular or made up, like humans, of many billions of cells divided into specialized tissues and organs.',
-      'vocabulary' => [
-        'name' => 'The Ontology for Biomedical Investigation',
-        'namespace' => 'obi',
-        'idspace' => 'OBI',
-        'description' => 'The Ontology for Biomedical Investigations (OBI) will serve as a resource for annotating biomedical investigations, including the study design, protocols and instrumentation used, the data generated and the types of analysis performed on the data.',
-        'url' => 'http://obi-ontology.org/page/Main_Page',
-        'urlprefix' => 'http://purl.obolibrary.org/obo/{db}_{accession}',
-      ],
-    ];
-
-    // Gene.
-    $terms['SO:0000704'] = [
-      'accession' => '0000704',
-      'name' => 'gene',
-      'definition' => 'A region (or regions) that includes all of the sequence elements necessary to encode a functional transcript. A gene may include regulatory regions, transcribed regions and/or other functional sequence regions.',
-      'vocabulary' => [
-        'name' => 'The Sequence Ontology',
-        'namespace' => 'sequence',
-        'idspace' => 'SO',
-        'description' => 'The Sequence Ontology (SO) is a collaborative ontology project for the definition of sequence features used in biological sequence annotation.',
-        'url' => 'http://www.sequenceontology.org/',
-        'urlprefix' => 'http://www.sequenceontology.org/browser/current_svn/term/{db}:{accession}',
-      ],
-    ];
-
-    // Germplasm Accession.
-    $terms['CO_010:0000044'] = [
-      'accession' => '0000044',
-      'name' => 'accession',
-      'definition' => '',
-      'vocabulary' => [
-        'name' => 'GCP germplasm ontology',
-        'namespace' => 'germplasm_ontology',
-        'idspace' => 'CO_010',
-        'description' => 'Provides desciptors for germplasm collections. Adapted from Descriptors for Banana (Musa spp.) (1996) and Descriptors for Mango by Bioversity.',
-        'url' => 'http://www.cropontology.org/ontology/CO_010/Germplasm',
-        'urlprefix' => 'http://www.cropontology.org/terms/{db}:{accession}/index.html',
-      ],
-    ];
-
-    // Analysis.
-    $terms['2945'] = [
-      'accession' => '2945',
-      'name' => 'Analysis',
-      'definition' => 'Apply analytical methods to existing data of a specific type.',
-      'vocabulary' => [
-        'name' => 'EDAM - Ontology of bioscientific data analysis',
-        'namespace' => 'EDAM',
-        'idspace' => 'operation',
-        'description' => 'EDAM is a comprehensive ontology of well-established, familiar concepts that are prevalent within computational biology, bioinformatics, and bioimage informatics.',
-        'url' => 'http://edamontology.org/page',
-        'urlprefix' => 'http://edamontology.org/{db}_{accession}',
-      ],
-    ];
-
-    /* Template for adding more.
-    $terms[''] = [
-      'accession' => '',
-      'name' => '',
-      'definition' => '',
-      'vocabulary' => [
-        'name' => '',
-        'namespace' => '',
-        'idspace' => '',
-        'description' => '',
-        'url' => '',
-        'urlprefix' => '',
-      ],
-    ]; */
-
-    // Actually add the terms described above.
-    // The term manager will create the vocabulary if it doesn't exist.
-    foreach ($terms as $term) {
-      \Drupal::service('tripal.tripalTerm.manager')->addTerm($term);
-    }
-
-    // ###################################################################
-    // Begin T3 code to be converted to D8 when the OBO importer is ready.
-    // ###################################################################
     /*
-      // Insert commonly used ontologies into the tables.
-      $ontologies = [
-        [
-          'name' => 'Relationship Ontology (legacy)',
-          'path' => '{tripal_chado}/files/legacy_ro.obo',
-          'auto_load' => FALSE,
-          'cv_name' => 'ro',
-          'db_name' => 'RO',
-        ],
-        [
-          'name' => 'Gene Ontology',
-          'path' => 'http://purl.obolibrary.org/obo/go.obo',
-          'auto_load' => FALSE,
-          'cv_name' => 'cellualar_component',
-          'db_name' => 'GO',
-        ],
-        [
-          'name' => 'Taxonomic Rank',
-          'path' => 'http://purl.obolibrary.org/obo/taxrank.obo',
-          'auto_load' => TRUE,
-          'cv_name' => 'taxonomic_rank',
-          'db_name' => 'TAXRANK',
-        ],
-        [
-          'name' => 'Tripal Contact',
-          'path' => '{tripal_chado}/files/tcontact.obo',
-          'auto_load' => TRUE,
-          'cv_name' => 'tripal_contact',
-          'db_name' => 'TContact',
-        ],
-        [
-          'name' => 'Tripal Publication',
-          'path' => '{tripal_chado}/files/tpub.obo',
-          'auto_load' => TRUE,
-          'cv_name' => 'tripal_pub',
-          'db_name' => 'TPUB',
-        ],
-        [
-          'name' => 'Sequence Ontology',
-          'path' => 'http://purl.obolibrary.org/obo/so.obo',
-          'auto_load' => TRUE,
-          'cv_name' => 'sequence',
-          'db_name' => 'SO',
-        ],
-      ];
+     This currently cannot be implementated as the vocabulary API is being
+     re-done. As such, this method is a placeholder.
 
-      module_load_include('inc', 'tripal_chado', 'includes/TripalImporter/OBOImporter');
-      for ($i = 0; $i < count($ontologies); $i++) {
-        $obo_id = chado_insert_obo($ontologies[$i]['name'], $ontologies[$i]['path']);
-        if ($ontologies[$i]['auto_load'] == TRUE) {
-          // Only load ontologies that are not already in the cv table.
-          $cv = chado_get_cv(['name' => $ontologies[$i]['cv_name']]);
-          $db = chado_get_db(['name' => $ontologies[$i]['db_name']]);
-          if (!$cv or !$db) {
-            print "Loading ontology: " . $ontologies[$i]['name'] . " ($obo_id)...\n";
-            $obo_importer = new OBOImporter();
-            $obo_importer->create(['obo_id' => $obo_id]);
-            $obo_importer->run();
-            $obo_importer->postRun();
-          }
-          else {
-            print "Ontology already loaded (skipping): " . $ontologies[$i]['name'] . "...\n";
-          }
-        }
-      }
+     See https://github.com/tripal/tripal/blob/7.x-3.x/tripal_chado/includes/setup/tripal_chado.setup.inc
+     for the Tripal 3 implementation of this method.
+
+     Vocabularies to be added individually:
+     - NCIT: NCI Thesaurus OBO Edition
+     - rdfs: Resource Description Framework Schema
+
+     Terms to be added individually:
+     - Subgroup (NCIT:C25693)
+     - rdfs:comment
+
+     Ontologies to be imported by the OBO Loader:
+     - Legacy Relationship Ontology: {tripal_chado}/files/legacy_ro.obo
+     - Gene Ontology: http://purl.obolibrary.org/obo/go.obo
+     - Taxonomic Rank: http://purl.obolibrary.org/obo/taxrank.obo
+     - Tripal Contact: {tripal_chado}/files/tcontact.obo
+     - Tripal Publication: {tripal_chado}/files/tpub.obo
+     - Sequence Ontology: http://purl.obolibrary.org/obo/so.obo
+     - Crop Ontology Germplasm: https://raw.githubusercontent.com/UofS-Pulse-Binfo/kp_entities/master/ontologies/CO_010.obo
+     - EDAM Ontology: http://edamontology.org/EDAM.obo
+
+     NOTE: Regarding CO_010 (crop ontology of germplasm), for some reason this
+     has been removed from the original crop ontology website. As such, I've linked
+     here to a file which loads and is correct. We use 4 terms from this ontology
+     for our content types so we may want to consider alternatives.
+     One such alternative may be MCPD: http://agroportal.lirmm.fr/ontologies/CO_020
+
     */
+
+    $this->logger->warning("\tWaiting on completion of the Vocabulary API and Data Loaders.");
   }
 
   /**
@@ -416,151 +253,580 @@ class ChadoPreparer extends ChadoTaskBase {
   }
 
   /**
+   * Helper: Create a given set of types.
+   *
+   * @param $types
+   *   An array of types to be created. The key is used to link the type to
+   *   it's term and the value is an array of details to be passed to the
+   *   create() method for Tripal Entity Types. Some keys should be:
+   *    - id: the integer id for this type; this will go away since it
+   *        should be automatic.
+   *    - name: the machine name for this type. It should be bio_data_[id]
+   *        where [id] matches the integer id above. Also should be automatic.
+   *    - label: a human-readable label for the content type.
+   *    - category: a grouping string to categorize content types in the UI.
+   *    - help_text: a single sentence describing the content type -usually
+   *        the default is the term definition.
+   * @param $terms
+   *   An array of terms which must already exist where the key maps to a
+   *   content type in the $types array. The value for each item is an array of
+   *   details to be passed to the term creation API. Some keys should be:
+   *    - accession: the unique identifier for the term (i.e. 2945)
+   *    - vocabulary:
+   *       - namespace: The name of the vocabulary (i.e. EDAM).
+   *       - idspace: the id space of the term (i.e. operation).
+   */
+  protected function createGivenContentTypes($types, $terms) {
+    foreach($terms as $key => $term_details) {
+      $type_details = $types[$key];
+
+      $this->logger->notice("  -- Creating " . $type_details['label'] . " (" . $type_details['name'] . ")...");
+
+      // TODO: Create the term once the API is upgraded.
+      // $term = \Drupal::service('tripal.tripalTerm.manager')->getTerms($term_details);
+
+      // TODO: Set the term in the type details.
+      if (is_object($term)) {
+        // $type_details['term_id'] = $term->getID();
+        if (!array_key_exists($type_details, 'help_text')) {
+          // $type_details['help_text'] = $term->getDefinition();
+        }
+      }
+      else {
+        $this->logger->warning("\tNo term attached -waiting on API update.");
+      }
+
+      // Check if the type already exists.
+      // TODO: use term instead of label once it's available.
+      $filter = ['label' => $type_details['label'] ];
+      $exists = \Drupal::entityTypeManager()
+        ->getStorage('tripal_entity_type')
+        ->loadByProperties($filter);
+
+      // Create the Type.
+      if (empty($exists)) {
+        $tripal_type = TripalEntityType::create($type_details);
+        if (is_object($tripal_type)) {
+          $tripal_type->save();
+          $this->logger->notice("\tSaved successfully.");
+        }
+        else {
+          $this->logger->error("\tCreation Failed! Details provided were: " . print_r($type_details));
+        }
+      }
+      else {
+        $this->logger->notice("\tSkipping as the content type already exists.");
+      }
+    }
+  }
+
+  /**
    * Creates the "General" category of content types.
+   *
+   * @code
+   $terms[''] =[
+     'accession' => '',
+     'vocabulary' => [
+       'idspace' => '',
+     ],
+   ];
+   $types['']= [
+     'id' => ,
+     'name' => '',
+     'label' => '',
+     'category' => 'General',
+   ];
+   * @endcode
    */
   protected function generalContentTypes() {
-    // Create the 'Organism' entity type. This uses the obi:organism term.
-    $term = \Drupal::service('tripal.tripalTerm.manager')->getTerms([
+
+    // The 'Organism' entity type. This uses the obi:organism term.
+    $terms['organism'] =[
       'accession' => '0100026',
       'vocabulary' => [
-        'namespace' => 'obi',
         'idspace' => 'OBI',
       ],
-    ]);
-
-    $organism = TripalEntityType::create([
+    ];
+    $types['organism']= [
       'id' => 1,
       'name' => 'bio_data_1',
       'label' => 'Organism',
-      'term_id' => $term->getID(),
-      'help_text' => $term->getDefinition(),
       'category' => 'General',
-    ]);
-    $organism->save();
+    ];
 
-    // Create the 'Analysis' entity type. This uses the EDAM:analysis term.
-    $term = \Drupal::service('tripal.tripalTerm.manager')->getTerms([
+    // The 'Analysis' entity type. This uses the EDAM:analysis term.
+    $terms['analysis'] = [
       'accession' => '2945',
       'vocabulary' => [
         'namespace' => 'EDAM',
         'idspace' => 'operation',
       ],
-    ]);
-
-    $analysis = TripalEntityType::create([
+    ];
+    $types['analysis'] = [
       'id' => 2,
       'name' => 'bio_data_2',
       'label' => 'Analysis',
-      'term_id' => $term->getID(),
-      'help_text' => $term->getDefinition(),
       'category' => 'General',
-    ]);
-    $analysis->save();
+    ];
 
-    // TODO: Create the 'Project' entity type. bio_data_3
+    // The 'Project' entity type. bio_data_3
+    $terms['project'] =[
+      'accession' => 'C47885',
+      'vocabulary' => [
+        'idspace' => 'NCIT',
+      ],
+    ];
+    $types['project']= [
+      'id' => 3,
+      'name' => 'bio_data_3',
+      'label' => 'Project',
+      'category' => 'General',
+    ];
 
-    // TODO: Create the 'Study' entity type. bio_data_4
+    // The 'Study' entity type. bio_data_4
+    $terms['study'] =[
+      'accession' => '001066',
+      'vocabulary' => [
+        'idspace' => 'SIO',
+      ],
+    ];
+    $types['study']= [
+      'id' => 4,
+      'name' => 'bio_data_4',
+      'label' => 'Study',
+      'category' => 'General',
+    ];
 
-    // TODO: Create the 'Contact' entity type. bio_data_5
+    // The 'Contact' entity type. bio_data_5
+    $terms['contact'] =[
+      'accession' => 'contact',
+      'vocabulary' => [
+        'idspace' => 'local',
+      ],
+    ];
+    $types['contact']= [
+      'id' => 5,
+      'name' => 'bio_data_5',
+      'label' => 'Contact',
+      'category' => 'General',
+    ];
 
-    // TODO: Create the 'Publication' entity type. bio_data_6
+    // The 'Publication' entity type. bio_data_6
+    $terms['publication'] =[
+      'accession' => '0000002',
+      'vocabulary' => [
+        'idspace' => 'TPUB',
+      ],
+    ];
+    $types['publication']= [
+      'id' => 6,
+      'name' => 'bio_data_6',
+      'label' => 'Publication',
+      'category' => 'General',
+    ];
 
-    // TODO: Create the 'Protocol' entity type. bio_data_7
+    // The 'Protocol' entity type. bio_data_7
+    $terms['protocol'] =[
+      'accession' => '00101',
+      'vocabulary' => [
+        'idspace' => 'sep',
+      ],
+    ];
+    $types['protocol']= [
+      'id' => 7,
+      'name' => 'bio_data_7',
+      'label' => 'Protocol',
+      'category' => 'General',
+    ];
+
+    $this->createGivenContentTypes($types, $terms);
   }
 
   /**
    * Creates the "Genomic" category of content types.
+   *
+   * @code
+   $terms[''] =[
+     'accession' => '',
+     'vocabulary' => [
+       'idspace' => '',
+     ],
+   ];
+   $types['']= [
+     'id' => ,
+     'name' => '',
+     'label' => '',
+     'category' => 'Genomic',
+   ];
+   * @endcode
    */
   protected function genomicContentTypes() {
-    // Create the 'Gene' entity type. This uses the sequence:gene term.
-    $term = \Drupal::service('tripal.tripalTerm.manager')->getTerms([
+
+    // The 'Gene' entity type. This uses the sequence:gene term.
+    $terms['gene'] = [
       'accession' => '0000704',
       'vocabulary' => [
         'namespace' => 'sequence',
         'idspace' => 'SO',
       ],
-    ]);
-
-    $gene = TripalEntityType::create([
+    ];
+    $types['gene'] = [
       'id' => 8,
       'name' => 'bio_data_8',
       'label' => 'Gene',
-      'term_id' => $term->getID(),
-      'help_text' => $term->getDefinition(),
       'category' => 'Genomic',
-    ]);
-    $gene->save();
+    ];
 
-    // TODO: Create the 'mRNA' entity type. bio_data_9
+    // the 'mRNA' entity type. bio_data_9
+    $terms['mRNA'] =[
+      'accession' => '0000234',
+      'vocabulary' => [
+        'idspace' => 'SO',
+      ],
+    ];
+    $types['mRNA']= [
+      'id' => 9,
+      'name' => 'bio_data_9',
+      'label' => 'mRNA',
+      'category' => 'Genomic',
+    ];
 
-    // TODO: Create the 'Phylogenetic tree' entity type. bio_data_10
+    // The 'Phylogenetic tree' entity type. bio_data_10
+    $terms['phylo'] =[
+      'accession' => '0872',
+      'vocabulary' => [
+        'idspace' => 'data',
+      ],
+    ];
+    $types['phylo']= [
+      'id' => 10,
+      'name' => 'bio_data_10',
+      'label' => 'Phylogenetic Tree',
+      'category' => 'Genomic',
+    ];
 
-    // TODO: Create the 'Physical Map' entity type. bio_data_11
+    // The 'Physical Map' entity type. bio_data_11
+    $terms['map'] =[
+      'accession' => '1280',
+      'vocabulary' => [
+        'idspace' => 'data',
+      ],
+    ];
+    $types['map']= [
+      'id' => 11,
+      'name' => 'bio_data_11',
+      'label' => 'Physical Map',
+      'category' => 'Genomic',
+    ];
 
-    // TODO: Create the 'DNA Library' entity type. bio_data_12
+    // The 'DNA Library' entity type. bio_data_12
+    $terms['library'] =[
+      'accession' => 'C16223',
+      'vocabulary' => [
+        'idspace' => 'NCIT',
+      ],
+    ];
+    $types['library']= [
+      'id' => 12,
+      'name' => 'bio_data_12',
+      'label' => 'DNA Library',
+      'category' => 'Genomic',
+    ];
 
-    // TODO: Create the 'Genome Assembly' entity type. bio_data_13
+    // The 'Genome Assembly' entity type. bio_data_13
+    $terms['assembly'] =[
+      'accession' => '0525',
+      'vocabulary' => [
+        'idspace' => 'operation',
+      ],
+    ];
+    $types['assembly']= [
+      'id' => 13,
+      'name' => 'bio_data_13',
+      'label' => 'Genome Assembly',
+      'category' => 'Genomic',
+    ];
 
-    // TODO: Create the 'Genome Annotation' entity type. bio_data_14
+    // The 'Genome Annotation' entity type. bio_data_14
+    $terms['annotation'] =[
+      'accession' => '0362',
+      'vocabulary' => [
+        'idspace' => 'operation',
+      ],
+    ];
+    $types['annotation']= [
+      'id' => 14,
+      'name' => 'bio_data_14',
+      'label' => 'Genome Assembly',
+      'category' => 'Genomic',
+    ];
 
-    // TODO: Create the 'Genome Project' entity type. bio_data_15
+    // The 'Genome Project' entity type. bio_data_15
+    $terms['genomeproject'] =[
+      'accession' => 'Genome Project',
+      'vocabulary' => [
+        'idspace' => 'local',
+      ],
+    ];
+    $types['genomeproject']= [
+      'id' => 15,
+      'name' => 'bio_data_15',
+      'label' => 'Genome Project',
+      'category' => 'Genomic',
+    ];
+
+    $this->createGivenContentTypes($types, $terms);
   }
 
   /**
    * Creates the "Genetic" category of content types.
+   *
+   * @code
+   $terms[''] =[
+     'accession' => '',
+     'vocabulary' => [
+       'idspace' => '',
+     ],
+   ];
+   $types['']= [
+     'id' => ,
+     'name' => '',
+     'label' => '',
+     'category' => 'Genetic',
+   ];
+   * @endcode
    */
   protected function geneticContentTypes() {
-    // TODO: Create the 'Genetic Map' entity type. bio_data_16
 
-    // TODO: Create the 'QTL' entity type. bio_data_17
+    // The 'Genetic Map' entity type. bio_data_16
+    $terms['map'] =[
+      'accession' => '1278',
+      'vocabulary' => [
+        'idspace' => 'data',
+      ],
+    ];
+    $types['map']= [
+      'id' => 16,
+      'name' => 'bio_data_16',
+      'label' => 'Genetic Map',
+      'category' => 'Genetic',
+    ];
 
-    // TODO: Create the 'Sequence Variant' entity type. bio_data_18
+    // The 'QTL' entity type. bio_data_17
+    $terms['qtl'] =[
+      'accession' => '0000771',
+      'vocabulary' => [
+        'idspace' => 'SO',
+      ],
+    ];
+    $types['qtl']= [
+      'id' => 17,
+      'name' => 'bio_data_17',
+      'label' => 'QTL',
+      'category' => 'Genetic',
+    ];
 
-    // TODO: Create the 'Genetic Marker' entity type. bio_data_19
+    // The 'Sequence Variant' entity type. bio_data_18
+    $terms['variant'] =[
+      'accession' => '0001060',
+      'vocabulary' => [
+        'idspace' => 'SO',
+      ],
+    ];
+    $types['variant']= [
+      'id' => 18,
+      'name' => 'bio_data_18',
+      'label' => 'Sequence Variant',
+      'category' => 'Genetic',
+    ];
 
-    // TODO: Create the 'Heritable Phenotypic Marker' entity type. bio_data_20
+    // The 'Genetic Marker' entity type. bio_data_19
+    $terms['marker'] =[
+      'accession' => '0001645',
+      'vocabulary' => [
+        'idspace' => 'SO',
+      ],
+    ];
+    $types['marker']= [
+      'id' => 19,
+      'name' => 'bio_data_19',
+      'label' => 'Genetic Marker',
+      'category' => 'Genetic',
+    ];
+
+    // The 'Heritable Phenotypic Marker' entity type. bio_data_20
+    $terms['hpn'] =[
+      'accession' => '0001500',
+      'vocabulary' => [
+        'idspace' => 'SO',
+      ],
+    ];
+    $types['hpn']= [
+      'id' => 20,
+      'name' => 'bio_data_20',
+      'label' => 'Heritable Phenotypic Marker',
+      'category' => 'Genetic',
+    ];
+
+    $this->createGivenContentTypes($types, $terms);
   }
 
   /**
-   * Creates the "Germplasm/Breeding" category of content types.
+   * Creates the "Germplasm" category of content types.
+   *
+   * @code
+   $terms[''] =[
+     'accession' => '',
+     'vocabulary' => [
+       'idspace' => '',
+     ],
+   ];
+   $types['']= [
+     'id' => ,
+     'name' => '',
+     'label' => '',
+     'category' => 'Germplasm',
+   ];
+   * @endcode
    */
   protected function germplasmContentTypes() {
-    // TODO: Create the 'Phenotypic Trait' entity type.
 
-    // Create the 'Germplasm Accession' entity type.
-    $term = \Drupal::service('tripal.tripalTerm.manager')->getTerms([
+    // The 'Phenotypic Trait' entity type. bio_data_28
+    $terms['trait'] =[
+      'accession' => 'C85496',
+      'vocabulary' => [
+        'idspace' => 'NCIT',
+      ],
+    ];
+    $types['trait']= [
+      'id' => 28,
+      'name' => 'bio_data_28',
+      'label' => 'Phenotypic Trait',
+      'category' => 'Germplasm',
+    ];
+
+    // The 'Germplasm Accession' entity type. bio_data_21
+    $terms['accession'] = [
       'accession' => '0000044',
       'vocabulary' => [
         'namespace' => 'germplasm_ontology',
         'idspace' => 'CO_010',
       ],
-    ]);
-
-    $germplasm_accession = TripalEntityType::create([
+    ];
+    $types['accession'] = [
       'id' => 21,
       'name' => 'bio_data_21',
       'label' => 'Germplasm Accession',
-      'term_id' => $term->getID(),
-      'help_text' => $term->getDefinition(),
-      'category' => 'Germplasm/Breeding',
-    ]);
-    $germplasm_accession->save();
+      'category' => 'Germplasm',
+    ];
 
-    // TODO: Create the 'Breeding Cross' entity type. bio_data_22
+    // The 'Breeding Cross' entity type. bio_data_22
+    $terms['cross'] =[
+      'accession' => '0000255',
+      'vocabulary' => [
+        'idspace' => 'CO_010',
+      ],
+    ];
+    $types['cross']= [
+      'id' => 22,
+      'name' => 'bio_data_22',
+      'label' => 'Breeding Cross',
+      'category' => 'Germplasm',
+    ];
 
-    // TODO: Create the 'Germplasm Variety' entity type. bio_data_23
+    // The 'Germplasm Variety' entity type. bio_data_23
+    $terms['variety'] =[
+      'accession' => '0000029',
+      'vocabulary' => [
+        'idspace' => 'CO_010',
+      ],
+    ];
+    $types['variety']= [
+      'id' => 23,
+      'name' => 'bio_data_23',
+      'label' => 'Germplasm Variety',
+      'category' => 'Germplasm',
+    ];
 
-    // TODO: Create the 'Recombinant Inbred Line' entity type. bio_data_24
+    // The 'Recombinant Inbred Line' entity type. bio_data_24
+    $terms['ril'] =[
+      'accession' => '0000162',
+      'vocabulary' => [
+        'idspace' => 'CO_010',
+      ],
+    ];
+    $types['ril']= [
+      'id' => 24,
+      'name' => 'bio_data_24',
+      'label' => 'Recombinant Inbred Line',
+      'category' => 'Germplasm',
+    ];
+
+    $this->createGivenContentTypes($types, $terms);
   }
 
   /**
    * Creates the "Expression" category of content types.
+   *
+   * @code
+   $terms[''] =[
+     'accession' => '',
+     'vocabulary' => [
+       'idspace' => '',
+     ],
+   ];
+   $types['']= [
+     'id' => ,
+     'name' => '',
+     'label' => '',
+     'category' => 'Expression',
+   ];
+   * @endcode
    */
   protected function expressionContentTypes() {
-    // TODO: Create the 'biological sample' entity type. bio_data_25
 
-    // TODO: Create the 'Assay' entity type. bio_data_26
+    // The 'biological sample' entity type. bio_data_25
+    $terms['sample'] =[
+      'accession' => '00195',
+      'vocabulary' => [
+        'idspace' => 'sep',
+      ],
+    ];
+    $types['sample']= [
+      'id' => 25,
+      'name' => 'bio_data_25',
+      'label' => 'Biological Sample',
+      'category' => 'Expression',
+    ];
 
-    // TODO: Create the 'Array Design' entity type. bio_data_27
+    // The 'Assay' entity type. bio_data_26
+    $terms['assay'] =[
+      'accession' => '0000070',
+      'vocabulary' => [
+        'idspace' => 'OBI',
+      ],
+    ];
+    $types['assay']= [
+      'id' => 26,
+      'name' => 'bio_data_26',
+      'label' => 'Assay',
+      'category' => 'Expression',
+    ];
+
+    // The 'Array Design' entity type. bio_data_27
+    $terms['design'] =[
+      'accession' => '0000269',
+      'vocabulary' => [
+        'idspace' => 'EFO',
+      ],
+    ];
+    $types['design']= [
+      'id' => 27,
+      'name' => 'bio_data_27',
+      'label' => 'Array Design',
+      'category' => 'Expression',
+    ];
+
+    $this->createGivenContentTypes($types, $terms);
   }
 }
